Refactor: Moved Reviews Routes to Separate File
Moved all review-related routes from the main server file to a separate reviews.js file for better organization.

Utilized express.Router with mergeParams: true to ensure that the :id parameter from the parent route (/campgrounds/:id/reviews) is accessible within the review routes.
