+++ app.js
@@ -11,7 +11,8 @@ const Review = require("./models/review");
 const methodOverride = require("method-override");
 const ejsMate = require("ejs-mate");
 
-const campgroundsRoute = require('./routes/campgrounds')
+const campgroundsRoute = require("./routes/campgrounds");
+const reviewsRoute = require("./routes/reviews");
 
 // use ejs-locals for all ejs templates:
 app.engine("ejs", ejsMate);
@@ -35,75 +36,25 @@ async function main() {
   console.log("Database is successfully connected.");
 }
 
-// Middleware to Validate a new campground using JOI
-// const validateCampground = (req, res, next) => {
-//   // use the JOI .validate method to get any errors
-//   // take the 'error' key from the resulting object after .validate(req.body)
-//   const { error } = campgroundSchema.validate(req.body);
-//   if (error) {
-//     // use map to iterate over each object in the 'details' array and extract the error message
-//     const msg = error.details.map((element) => element.message).join(",");
-//     // console.log(msg);
-//     throw new ExpressError(msg, 400);
-//   } else {
-//     next();
-//   }
-// };
-
 const validateReview = (req, res, next) => {
   const { error } = reviewSchema.validate(req.body);
 
   if (error) {
-    const msg = error.details.map((element) => element.message).join(",")
-    throw new ExpressError(msg, 400)
+    const msg = error.details.map((element) => element.message).join(",");
+    throw new ExpressError(msg, 400);
   } else {
     next();
   }
-}
+};
 
-app.use('/campgrounds', campgroundsRoute)
+// Use the Router objects created in the routes folder
+app.use("/campgrounds", campgroundsRoute);
+app.use("/campgrounds/:id/reviews", reviewsRoute);
 
 // ROUTES
 app.get("/", (req, res) => {
   res.render("home");
 });
-
-// allow user to delete a review
-app.delete(
-  "/campgrounds/:id/reviews/:reviewId",
-  catchAsync(async (req, res) => {
-    // find the campground and delete the corresponding id
-    // find the review using the id
-    const { id, reviewId } = req.params;
-
-    // using the specific campground to delete the specified review associated with it in the db
-    const removeCampgroundReview = await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
-
-    // delete the review from the reviews db collection
-    const deletedReview = await Review.findByIdAndDelete(reviewId);
-    // console.log(`This is the deleted review: ${deletedReview}`);
-    res.redirect(`/campgrounds/${id}`);
-  })
-);
-
-app.post(
-  "/campgrounds/:id/reviews",
-  validateReview,
-  catchAsync(async (req, res) => {
-    console.log(req.params);
-    const { id } = req.params;
-    const campground = await Campground.findById(id);
-
-    const review = new Review(req.body.review);
-
-    campground.reviews.push(review);
-
-    await review.save();
-    await campground.save();
-
-    res.redirect(`/campgrounds/${campground._id}`);
-  })
-);
 
 // Catch all for routes that don't match above routes
 app.all("*", (req, res, next) => {
