+++ routes/campgrounds.js
@@ -21,16 +21,6 @@ const validateCampground = (req, res, next) => {
     }
 };
 
-// const validateReview = (req, res, next) => {
-//     const { error } = reviewSchema.validate(req.body);
-
-//     if (error) {
-//         const msg = error.details.map((element) => element.message).join(",")
-//         throw new ExpressError(msg, 400)
-//     } else {
-//         next();
-//     }
-// }
 
 router.get(
     "/",
