+++ routes/reviews.js
@@ -1,0 +1,60 @@
+const express = require('express');
+const router = express.Router({ mergeParams: true });
+
+const catchAsync = require("../utilities/catchAsync");
+const ExpressError = require("../utilities/ExpressError");
+const Campground = require("../models/campground")
+const Review = require("../models/review")
+
+const { reviewSchema } = require("../schemas");
+
+
+// Middleware to validate a review
+const validateReview = (req, res, next) => {
+    const { error } = reviewSchema.validate(req.body);
+
+    if (error) {
+        const msg = error.details.map((element) => element.message).join(",")
+        throw new ExpressError(msg, 400)
+    } else {
+        next();
+    }
+}
+
+// allow a user to post a review
+router.post("/", validateReview, catchAsync(async (req, res) => {
+    // console.log(req.params);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+
+    const review = new Review(req.body.review);
+
+    console.log(`Campground: ${campground}`)
+
+    campground.reviews.push(review);
+
+    await review.save();
+    await campground.save();
+
+    res.redirect(`/campgrounds/${campground._id}`);
+})
+);
+
+
+// allow user to delete a review
+router.delete("/:reviewId", catchAsync(async (req, res) => {
+    // find the campground and delete the corresponding id
+    // find the review using the id
+    const { id, reviewId } = req.params;
+
+    // using the specific campground to delete the specified review associated with it in the db
+    const removeCampgroundReview = await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
+
+    // delete the review from the reviews db collection
+    const deletedReview = await Review.findByIdAndDelete(reviewId);
+    // console.log(`This is the deleted review: ${deletedReview}`);
+    res.redirect(`/campgrounds/${id}`);
+})
+);
+
+module.exports = router;
